Slightly Changed home.go
Main reason for change was a bug that was solved because  s (the SMTPInfo) was created as a pointer for some reason and was getting a nil error when data was sent to it. The two defer statements that have warnings are required and will break the  upload of the second file upload.